+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 2020-02-26
 
-* TODO: Describe initial release.
+UIC (UI Components) is a set of Flutter widgets that simplifies implementing most used UI cases.
+
+Each UIC widget uses standard Flutter widgets to implements its view, but adds additional logic, like state management, and parameters to customize its appearance.
+
+Here is the list of available components:
+- `ListUic` - Wrapper of ListView, which implements related data loading and state management logic.
+- `ProgressUic` - Wrapper of ProgressIndicator with additional text.